fix(worker): extract exit_diagnostics into its own module (#268)
CI failed to collect ``tests/unit/test_worker_exit_diagnostics.py``
because importing :mod:`aios.harness.worker` transitively imports
:mod:`aios.harness.procrastinate_app`, which constructs a
``procrastinate.App`` at module load and calls ``get_settings()``
inside that constructor.  Without a ``.env`` file (CI runners), the
``Settings`` validator raises before the test module is even
collected, and pytest's ``_unit_env`` autouse fixture only fires
post-collection.

Moving :func:`install_exit_diagnostics` into a new
:mod:`aios.harness.exit_diagnostics` module with no aios-side imports
breaks the chain: tests import from the small module directly, the
worker imports from it as before.  No behavior change.
